Add cli_actions.py
